+++ lib/mcp/tools/tools.dart
@@ -2,18 +2,17 @@
 library;
 
 import 'base_tool.dart';
-import 'calculator_tool.dart';
-import 'pubdev_tool.dart';
-import 'url_fetch_tool.dart';
-import 'wikipedia_tool.dart';
+import 'evaluate_expression.dart';
+import 'fetch_url_tool.dart';
+import 'pubdev_search_tool.dart';
+import 'wikipedia_search_tool.dart';
 
 export 'base_tool.dart';
-export 'calculator_tool.dart';
 
 /// Creates all available tools.
 List<BaseTool> createAllTools() => [
-      CalculatorTool(),
       UrlFetchTool(),
-      PubDevSearchTool(),
       WikipediaSearchTool(),
+      EvaluateExpressionTool(),
+      PubDevSearchTool(),
     ];